+++ src/Database/Beam/MySQL/FromField.hs
@@ -1,19 +1,35 @@
+{-# LANGUAGE BangPatterns        #-}
 {-# LANGUAGE MultiWayIf          #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Database.Beam.MySQL.FromField
     ( FieldParser
     , FromField(..)) where
-    -- , atto ) where
 
+import           Control.Applicative               ((<|>))
 import           Control.Monad.Except              (ExceptT, throwError)
-import           Data.Attoparsec.ByteString.Char8  (Parser, decimal, double,
-                                                    parseOnly, signed)
+import           Data.Aeson                        (Value, eitherDecode')
+import           Data.Attoparsec.ByteString.Char8  (Parser, char, decimal,
+                                                    digit, double, parseOnly,
+                                                    rational, signed,
+                                                    takeByteString,
+                                                    takeLazyByteString)
 import           Data.Bits                         (finiteBitSize, zeroBits)
+import qualified Data.ByteString                   as BS
 import           Data.ByteString.Char8             (ByteString, uncons)
+import qualified Data.ByteString.Lazy              as BSL
+import           Data.Char                         (digitToInt)
 import           Data.Int                          (Int16, Int32, Int64, Int8)
+import           Data.Scientific                   (Scientific)
+import           Data.Text                         (Text)
+import           Data.Text.Encoding                (decodeUtf8')
+import           Data.Time.Calendar                (Day, fromGregorianValid)
+import           Data.Time.Clock                   (NominalDiffTime)
+import           Data.Time.LocalTime               (LocalTime (..), TimeOfDay,
+                                                    makeTimeOfDayValid)
 import           Data.Word                         (Word16, Word32, Word64,
                                                     Word8)
+import           Database.Beam.Backend.SQL         (SqlNull (..))
 import           Database.Beam.Backend.SQL.Row     (ColumnParseError (..))
 import           Database.MySQL.Protocol.ColumnDef (ColumnDef (columnType),
                                                     FieldType, mySQLTypeBit,
@@ -122,6 +138,72 @@ instance FromField Double where
            | t == mySQLTypeNewDecimal -> True
            | otherwise -> False
 
+instance FromField Scientific where
+  fromField = parseChecking "Scientific" representableScientific rational
+
+instance FromField Rational where
+  fromField = parseChecking "Rational" representableScientific rational
+
+instance (FromField a) => FromField (Maybe a) where
+  fromField f = \case
+    Nothing -> pure Nothing
+    d -> Just <$> fromField f d
+
+instance FromField SqlNull where
+  fromField f = \case
+    Nothing -> pure SqlNull
+    _ -> do
+      let st = fieldTypeToString . columnType $ f
+      throwError (ColumnTypeMismatch "SqlNull" st "Non-null value found")
+
+instance FromField BS.ByteString where
+  fromField = parseChecking "ByteString" representableBytes takeByteString
+
+instance FromField BSL.ByteString where
+  fromField = parseChecking "ByteString.Lazy" representableBytes takeLazyByteString
+
+instance FromField Text where
+  fromField = parseChecking "Text" representableText (go =<< takeByteString)
+    where
+      go bs = case decodeUtf8' bs of
+        Left err -> fail . show $ err
+        Right t  -> pure t
+
+instance FromField LocalTime where
+  fromField = parseChecking "LocalTime" check p
+    where
+      p = do
+        (day, time) <- parseDayAndTime
+        pure (LocalTime day time)
+      check t =
+        if | t == mySQLTypeDateTime -> True
+           | t == mySQLTypeTimestamp -> True
+           | t == mySQLTypeDate -> True
+           | otherwise -> False
+
+instance FromField Day where
+  fromField = parseChecking "Day" (mySQLTypeDate ==) parseDay
+
+instance FromField TimeOfDay where
+  fromField = parseChecking "TimeOfDay" (mySQLTypeTime ==) parseTime
+
+instance FromField NominalDiffTime where
+  fromField = parseChecking "NominalDiffTime" (mySQLTypeTime ==) p
+    where
+      p = do
+        negative <- (True <$ char '-') <|> pure False
+        hours <- lengthedDecimal 3 <|> lengthedDecimal 2
+        (minutes, seconds, microseconds) <- parseMSU
+        let delta = hours * 3600 + minutes * 60 + seconds + microseconds * 1e-6
+        pure (if negative then negate delta else delta)
+
+instance FromField Value where
+  fromField = parseChecking "Aeson.Value" representableBytes (go =<< takeLazyByteString)
+    where
+      go lbs = case eitherDecode' lbs of
+        Left err -> fail err
+        Right v  -> pure v
+
 -- Helpers
 
 -- HOF to automate the parsing of many common types
@@ -171,6 +253,100 @@ fitsWord =
     then fits64
     else fits32 -- conservative, but better safe than sorry
 
+-- Checks if we can represent this a a blob of bytes
+representableBytes :: FieldType -> Bool
+representableBytes t =
+  if | representableText t -> True
+     | t == mySQLTypeTinyBlob -> True
+     | t == mySQLTypeMediumBlob -> True
+     | t == mySQLTypeLongBlob -> True
+     | otherwise -> False
+
+-- Checks if we can represent this as text
+representableText :: FieldType -> Bool
+representableText t =
+  if | t == mySQLTypeVarChar -> True
+     | t == mySQLTypeVarString -> True
+     | t == mySQLTypeString -> True
+     | t == mySQLTypeEnum -> True
+     | t == mySQLTypeBlob -> True -- suspicious
+     | otherwise -> False
+
+-- Checks if a type can be represented as a Scientific
+representableScientific :: FieldType -> Bool
+representableScientific t =
+  if | fits64 t -> True
+     | t == mySQLTypeFloat -> True
+     | t == mySQLTypeDouble -> True
+     | t == mySQLTypeDecimal -> True
+     | t == mySQLTypeNewDecimal -> True
+     | otherwise -> False
+
+parseDayAndTime :: Parser (Day, TimeOfDay)
+parseDayAndTime = do
+  day <- parseDay
+  _ <- char ' '
+  time <- parseTime
+  pure (day, time)
+
+-- Attempts to parse HH:MM:SS.MMMMMM format timestamp
+parseTime :: Parser TimeOfDay
+parseTime = do
+  hours <- lengthedDecimal 2
+  _ <- char ':'
+  (minutes, seconds, microseconds) <- parseMSU
+  let pico = seconds + microseconds * 1e-6
+  case makeTimeOfDayValid hours minutes pico of
+    Nothing  -> fail (msg hours minutes seconds microseconds)
+    Just tod -> pure tod
+  where
+    msg h m s u =
+      "Invalid time: " <> show h <> ":" <> show m <> ":" <> show s <> "." <> show u
+
+-- Attempts to parse ':MM:SS.MMMMMM'-style segment
+parseMSU :: (Num a, Num b, Num c) => Parser (a, b, c)
+parseMSU = do
+  _ <- char ':'
+  m <- lengthedDecimal 2
+  _ <- char ':'
+  s <- lengthedDecimal 2
+  u <- (char '.' *> maxLengthedDecimal 6) <|> pure 0
+  pure (m, s, u)
+
+-- Attempts to parse YYYY-MM-DD format datestamp
+parseDay :: Parser Day
+parseDay = do
+  year <- lengthedDecimal 4
+  _ <- char '-'
+  month <- lengthedDecimal 2
+  _ <- char '-'
+  day <- lengthedDecimal 2
+  case fromGregorianValid year month day of
+    Nothing   -> fail (msg year month day)
+    Just day' -> pure day'
+  where
+    msg y m d =
+      "Invalid date: " <> show y <> "-" <> show m <> "-" <> show d
+
+-- Parse a decimal with exactly n digits
+lengthedDecimal :: (Num a) => Word -> Parser a
+lengthedDecimal = go' 0
+  where
+    go' !x 0 = pure x
+    go' !x n = do
+      d <- digitToInt <$> digit
+      go' (x * 10 + fromIntegral d) (n - 1)
+
+-- Parse a decimal with at most n digits
+maxLengthedDecimal :: (Num a) => Word -> Parser a
+maxLengthedDecimal = go 0
+  where
+    go x n = do
+      d <- digitToInt <$> digit
+      go' (x * 10 + fromIntegral d) (n - 1)
+    go' !x 0 = pure x
+    go' !x n = go x n <|> pure (x * 10 ^ n)
+
 -- Stringification of type names
 fieldTypeToString :: FieldType -> String
 fieldTypeToString ft = "MySQL " <>
@@ -204,271 +380,3 @@ fieldTypeToString ft = "MySQL " <>
      | ft == mySQLTypeVarString -> "VarString"
      | ft == mySQLTypeString -> "String"
      | otherwise -> "Geometry" -- brittle, to fix
-
-{-
-instance FromField Bool where
-    fromField f Nothing     = unexpectedNull f ""
-    fromField f md@(Just d) = (/= 0) <$>
-        case fieldType f of
-          Tiny -> fromField f md
-          Bit  -> case parseBit d of
-            Left err -> conversionFailed f err
-            Right r  -> pure r
-          _    -> incompatibleTypes f ""
-      where
-        parseBit :: SB.ByteString -> Either String Word8
-        parseBit bs
-          | BS.null bs = Left  (   show bs)
-          | otherwise  = Right (BS.head bs)
-
-instance FromField Word where
-    fromField = atto check64 decimal
-instance FromField Word8 where
-    fromField = atto check8 decimal
-instance FromField Word16 where
-    fromField = atto check16 decimal
-instance FromField Word32 where
-    fromField = atto check32 decimal
-instance FromField Word64 where
-    fromField = atto check64 decimal
-
-instance FromField Int where
-    fromField = atto check64 (signed decimal)
-instance FromField Int8 where
-    fromField = atto check8 (signed decimal)
-instance FromField Int16 where
-    fromField = atto check16 (signed decimal)
-instance FromField Int32 where
-    fromField = atto check32 (signed decimal)
-instance FromField Int64 where
-    fromField = atto check64 (signed decimal)
-
-instance FromField Float where
-    fromField = atto check (realToFrac <$> double)
-      where
-        check ty         | check16 ty = True
-        check Int24      = True
-        check Float      = True
-        check Decimal    = True
-        check NewDecimal = True
-        check Double     = True
-        check _=         False
-
-instance FromField Double where
-    fromField = atto check double
-      where
-        check ty         | check32 ty = True
-        check Float      = True
-        check Double     = True
-        check Decimal    = True
-        check NewDecimal = True
-        check _          = False
-
-instance FromField Scientific where
-    fromField = atto checkScientific rational
-
-instance FromField (Ratio Integer) where
-    fromField = atto checkScientific rational
-
-instance FromField a => FromField (Maybe a) where
-    fromField _ Nothing      = pure Nothing
-    fromField field (Just d) = Just <$> fromField field (Just d)
-
-instance FromField SqlNull where
-    fromField _ Nothing = pure SqlNull
-    fromField f _ = throwError (ColumnTypeMismatch "SqlNull"
-                                                   (show (fieldType f))
-                                                   "Non-null value found")
-
-instance FromField SB.ByteString where
-    fromField = doConvert checkBytes pure
-
-instance FromField LB.ByteString where
-    fromField f d = fmap (LB.fromChunks . pure) (fromField f d)
-
-instance FromField TS.Text where
-    fromField = doConvert checkText (either (Left . show) Right . TE.decodeUtf8')
-
-instance FromField TL.Text where
-    fromField f d = fmap (TL.fromChunks . pure) (fromField f d)
-
-instance FromField LocalTime where
-    fromField = atto checkDate localTime
-      where
-        checkDate DateTime  = True
-        checkDate Timestamp = True
-        checkDate Date      = True
-        checkDate _         = False
-
-        localTime = do
-          (day, time) <- dayAndTime
-          pure (LocalTime day time)
-
-instance FromField Day where
-    fromField = atto checkDay dayP
-      where
-        checkDay Date = True
-        checkDay _    = False
-
-instance FromField TimeOfDay where
-    fromField = atto checkTime timeP
-      where
-        checkTime Time = True
-        checkTime _    = False
-
-instance FromField NominalDiffTime where
-    fromField = atto checkTime durationP
-      where
-        checkTime Time = True
-        checkTime _    = False
-
-instance FromField A.Value where
-    fromField f bs =
-        case (maybeToRight "Failed to extract JSON bytes." bs) >>= A.eitherDecodeStrict of
-            Left err -> conversionFailed f err
-            Right x  -> pure x
-
-dayAndTime :: Parser (Day, TimeOfDay)
-dayAndTime = do
-  day <- dayP
-  _ <- char ' '
-  time <- timeP
-
-  pure (day, time)
-
-timeP :: Parser TimeOfDay
-timeP = do
-  hour <- lengthedDecimal 2
-  _ <- char ':'
-  minute <- lengthedDecimal 2
-  _ <- char ':'
-  seconds <- lengthedDecimal 2
-  microseconds <- (char '.' *> maxLengthedDecimal 6) <|>
-                  pure 0
-
-  let pico = seconds + microseconds * 1e-6
-  case makeTimeOfDayValid hour minute pico of
-    Nothing -> fail (printf "Invalid time part: %02d:%02d:%s" hour minute (showFixed False pico))
-    Just tod -> pure tod
-
-durationP :: Parser NominalDiffTime
-durationP = do
-  negative <- (True <$ char '-') <|> pure False
-  hour <- lengthedDecimal 3
-  _ <- char ':'
-  minute <- lengthedDecimal 2
-  _ <- char ':'
-  seconds <- lengthedDecimal 2
-  microseconds <- (char '.' *> maxLengthedDecimal 6) <|>
-                  pure 0
-
-  let v = hour * 3600 + minute * 60 + seconds +
-          microseconds * 1e-6
-
-  pure (if negative then negate v else v)
-
-dayP :: Parser Day
-dayP = do
-  year <- lengthedDecimal 4
-  _ <- char '-'
-  month <- lengthedDecimal 2
-  _ <- char '-'
-  day <- lengthedDecimal 2
-
-  case fromGregorianValid year month day of
-    Nothing -> fail (printf "Invalid date part: %04d-%02d-%02d" year month day)
-    Just day' -> pure day'
-
-lengthedDecimal :: Num a => Int -> Parser a
-lengthedDecimal = lengthedDecimal' 0
-  where
-    lengthedDecimal' !a 0 = pure a
-    lengthedDecimal' !a n = do
-      d <- digitToInt <$> digit
-      lengthedDecimal' (a * 10 + fromIntegral d) (n - 1)
-
-maxLengthedDecimal :: Num a => Int -> Parser a
-maxLengthedDecimal = go1 0
-  where
-    go1 a n = do
-      d <- digitToInt <$> digit
-      go' (a * 10 + fromIntegral d) (n - 1)
-
-    go' !a 0 = pure a
-    go' !a n =
-      go1 a n <|> pure (a * 10  ^ n)
-
-incompatibleTypes, unexpectedNull, conversionFailed
-    :: forall a. Typeable a => Field -> String -> FieldParser a
-incompatibleTypes f msg =
-  throwError (ColumnTypeMismatch (show (typeRep (Proxy :: Proxy a)))
-                                 (show (fieldType f))
-                                 msg)
-unexpectedNull _ _ =
-  throwError ColumnUnexpectedNull
-conversionFailed f msg =
-  throwError (ColumnTypeMismatch (show (typeRep (Proxy :: Proxy a)))
-                                 (show (fieldType f))
-                                 msg)
-
-check8, check16, check32, check64, checkScientific, checkBytes, checkText
-    :: Type -> Bool
-
-check8 Tiny       = True
-check8 NewDecimal = True
-check8 _          = False
-
-check16 ty    | check8 ty = True
-check16 Short = True
-check16 _=    False
-
-check32 ty    | check16 ty = True
-check32 Int24 = True
-check32 Long  = True
-check32 _     = False
-
-check64 ty       | check32 ty = True
-check64 LongLong = True
-check64 _        = False
-
-checkScientific ty         | check64 ty = True
-checkScientific Float      = True
-checkScientific Double     = True
-checkScientific Decimal    = True
-checkScientific NewDecimal = True
-checkScientific _          = True
-
-checkBytes ty         | checkText ty = True
-checkBytes TinyBlob   = True
-checkBytes MediumBlob = True
-checkBytes LongBlob   = True
-checkBytes Blob       = True
-checkBytes _          = False
-
-checkText VarChar   = True
-checkText VarString = True
-checkText String    = True
-checkText Enum      = True
-checkText Blob      = True
-checkText _         = False
-
-doConvert :: Typeable a => (Type -> Bool)
-          -> (SB.ByteString -> Either String a)
-          -> Field -> Maybe SB.ByteString -> FieldParser a
-doConvert _ _ f Nothing = unexpectedNull f ""
-doConvert checkType parser field (Just d)
-  | checkType (fieldType field) =
-      case parser d of
-        Left err -> conversionFailed field err
-        Right r  -> pure r
-  | otherwise = incompatibleTypes field ""
-
-atto :: Typeable a => (Type -> Bool) -> Parser a
-     -> Field -> Maybe SB.ByteString -> FieldParser a
-atto checkType parser =
-  doConvert checkType (parseOnly parser)
-
-maybeToRight :: b -> Maybe a -> Either b a
-maybeToRight _ (Just x) = Right x
-maybeToRight y Nothing  = Left y
--}